--- cpp/open3d/utility/OMPProgressBar.h
@@ -1,54 +1,0 @@
-// ----------------------------------------------------------------------------
-// -                        Open3D: www.open3d.org                            -
-// ----------------------------------------------------------------------------
-// The MIT License (MIT)
-//
-// Copyright (c) 2018-2021 www.open3d.org
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files (the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions:
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
-// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
-// IN THE SOFTWARE.
-// ----------------------------------------------------------------------------
-
-#pragma once
-
-#ifdef _OPENMP
-/// Multi-threading - yay!
-#include <omp.h>
-#else
-/// Macros used to disguise the fact that we do not have multithreading enabled.
-#define omp_get_thread_num() 0
-#define omp_get_num_threads() 1
-#endif
-
-#include "open3d/utility/ProgressBar.h"
-
-namespace open3d {
-namespace utility {
-
-class OMPProgressBar : public ProgressBar {
-public:
-    OMPProgressBar(size_t expected_count,
-                   const std::string &progress_info,
-                   bool active = false)
-        : ProgressBar(expected_count, progress_info, active){};
-
-    ProgressBar &operator++() override;
-};
-
-}  // namespace utility
-}  // namespace open3d
