+++ cpp/open3d/utility/ProgressBar.cpp
@@ -28,6 +28,15 @@
 
 #include <fmt/printf.h>
 
+#ifdef _OPENMP
+/// Multi-threading - yay!
+#include <omp.h>
+#else
+/// Macros used to disguise the fact that we do not have multithreading enabled.
+#define omp_get_thread_num() 0
+#define omp_get_num_threads() 1
+#endif
+
 namespace open3d {
 namespace utility {
 
@@ -81,6 +90,23 @@ void ProgressBar::UpdateCurrentCount(size_t n) {
 }
 
 size_t ProgressBar::GetCurrentCount() const { return current_count_; }
+
+ProgressBar &OMPProgressBar::operator++() {
+    // Ref: https://stackoverflow.com/a/44555438
+
+    int number_of_threads = omp_get_num_threads();
+    int thread_id = omp_get_thread_num();
+
+    // check if inside OMP loop
+    if (number_of_threads > 1) {
+        // check thread_id == 0
+        if (thread_id == 0) UpdateCurrentCount(number_of_threads);
+    } else {
+        UpdateCurrentCount(number_of_threads);
+    }
+
+    return *this;
+}
 
 }  // namespace utility
 }  // namespace open3d
