+++ cpp/open3d/utility/ProgressBar.h
@@ -36,26 +36,31 @@ public:
     ProgressBar(size_t expected_count,
                 const std::string &progress_info,
                 bool active = false);
-
     void Reset(size_t expected_count,
                const std::string &progress_info,
                bool active);
-
     virtual ProgressBar &operator++();
-
     void SetCurrentCount(size_t n);
-
     void UpdateCurrentCount(size_t n);
-
     size_t GetCurrentCount() const;
 
-private:
+protected:
     const size_t resolution_ = 40;
     size_t expected_count_;
     size_t current_count_;
     std::string progress_info_;
     size_t progress_pixel_;
     bool active_;
+};
+
+class OMPProgressBar : public ProgressBar {
+public:
+    OMPProgressBar(size_t expected_count,
+                   const std::string &progress_info,
+                   bool active = false)
+        : ProgressBar(expected_count, progress_info, active) {}
+
+    ProgressBar &operator++() override;
 };
 
 }  // namespace utility
